+++ src/main/java/JUnitTests/TestAppDatabase.java
@@ -1,4 +1,4 @@
-package Database;
+package JUnitTests;
 
 import Database.AppDatabase;
 import Database.Record;
