+++ src/main/java/JUnitTests/TestBinarySearchTreeRunner.java
@@ -1,4 +1,4 @@
-package BinarySearchTree;
+package JUnitTests;
 
 import org.junit.runner.JUnitCore;
 import org.junit.runner.Result;
